requirements.txt
numpy==1.19.5
flask==1.1.1
keras-preprocessing==1.1.2
gunicorn